Check if lock file exists to avoid warnings.

+++ cron/lib/shutdown.php
@@ -3,7 +3,8 @@ register_shutdown_function('shutdown');
 function shutdown() {
 	global $lock_file;
 	
-	unlink($lock_file);
+	if (file_exists($lock_file))
+		unlink($lock_file);
 }
 
 ?>
